+++ surpyval/univariate/competing_risks/__init__.py
@@ -1,8 +1,10 @@
 from .nonparametric import CompetingRisks
+from .parametric import ParametricCompetingRisks
 from .regression import CompetingRisksProportionalHazards, FineGray
 
 __all__ = [
     "CompetingRisks",
+    "ParametricCompetingRisks",
     "CompetingRisksProportionalHazards",
     "FineGray",
 ]