+++ surpyval/univariate/competing_risks/parametric/__init__.py
@@ -1,0 +1,3 @@
+from .parametric_competing_risks import ParametricCompetingRisks
+
+__all__ = ["ParametricCompetingRisks"]